backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.22.1
psycopg[binary]
alembic==1.13.0
pydantic>=2.12.5
pydantic-settings==2.5.0
PyJWT==2.12.0
bcrypt==4.2.0
python-multipart==0.0.22
httpx==0.27.0
openai==1.50.0
aiofiles==24.1.0
websockets==13.1
pytest==8.3.5
redis==5.2.1
